phase 4a of issue #82: split processing.py into processing/ package
